Simple space page for invite token
