+++ hub/static/hub/base.css
@@ -40,6 +40,8 @@ header {
     font-weight: 800;
     max-height: 3.2rem;
     border: none;
+    overflow: hidden;
+    text-overflow: ellipsis;
 }
 
 .button:link {
@@ -118,27 +120,17 @@ label {
     margin-right: auto;
 }
 
-.space-dropdown {
-    position: relative;
-    display: inline-block;
+
+.horizontal-line {
+    width: 100%;
+    height: 2px;
+    background-color: var(--light-bg);
 }
 
-.quick-space-container {
-    position: absolute;
-    display: none;
-    min-width: 15vw;
-    top: 3rem;
-    background-color: var(--mid-bg);
-    padding: 1rem;
-    right: 0;
-
-    flex-direction: column;
-    align-items: center;
-    gap: 0.5rem;
-}
-
-.space-row {
+.center-container {
+    width: 100%;
     display: flex;
     flex-direction: row;
-    gap: 0.25rem;
+    align-items: center;
+    justify-content: center;
 }
