+++ hub/templates/hub/base.html
@@ -5,6 +5,7 @@
 {% load django_htmx %}
 <head>
     <link rel="stylesheet" href="{% static 'hub/base.css' %}">
+    <link rel="stylesheet" href="{% static 'space/space.css' %}">
     <title>{% block title %}Household Planner{% endblock %}</title>
 
     <link rel="preconnect" href="https://fonts.bunny.net">
@@ -33,7 +34,7 @@
         <div class="header-section">
             <div class="space-dropdown">
                 <a class="button space-button" onclick="toggleQuickspace(event)">Spaces</a>
-                {% include "hub/components/quick_space.html" %}
+                {% include "space/components/quick_space.html" %}
             </div>
             
             <a href="{% url 'hub:logout' %}" class="button">Logout</a>
