+++ root/urls.py
@@ -24,5 +24,6 @@ urlpatterns = [
     path('transactions/', include(("transactions.urls", "transactions"), namespace="transactions")),
     path('todos/', include(("todos.urls", "urls"), namespace="todos")),
     path('', include(("hub.urls", "hub"), namespace="hub")),
+    path('space/', include(("space.urls", "space"), namespace="space")),
 
 ]
