+++ space/admin.py
@@ -1,0 +1,6 @@
+from django.contrib import admin
+
+from .models import SharedSpace
+
+# Register your models here.
+admin.register(SharedSpace)
